full review

+++ oracle/m4/s10.sql
@@ -36,7 +36,7 @@ alter table execs add constraint execs_name_uq unique(first_name, last_name);
 insert into execs values(1101, 'Neena', 'Kochhar', sysdate, 15000.00);
 
 -- add pk
-alter table execs modify exec_id number(6) primary key;
+alter table execs modify exec_id integer primary key;
 
 -- drop pk
 -- alter table execs drop primary key;
